requirements.txt
# Core libraries
python-frontmatter==1.0.0
markdown==3.4.3
asyncio==3.4.3
aiohttp==3.8.5
pyyaml==6.0
sentence-transformers==2.2.2
pinecone-client==2.2.2
sqlalchemy==2.0.20

# Utilities
psutil==5.9.5
tqdm==4.66.1
colorlog==6.7.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1